Scaled the scene up 1.3x + additions
Changed the scale of the caves only, the gate area still same size. The origin cube is still at 0,0,0 but the level has moved because of the scaling.
The original 0,0,0 cube has been lost because of old model deletion, but the level's pivot is still at 0,0,0.
Changed lighting to mixed, this means direct light is realtime, this may be a horrid mistake for mobile, but I want to see if it's possible because the visual upgrades are nice.

Added great cave paintings.